Make Alembic upgrade work if DB inconsistent
The Weasyl development VM is missing tables which have not been removed with an Alembic script. As such, 'alembic upgrade head' fails. Therefore, manually check for the tables, and purge if required.

+++ libweasyl/libweasyl/alembic/versions/00222d3a3142_partial_schema_cleanup_purges_tables_.py
@@ -10,32 +10,48 @@ Create Date: 2016-06-30 02:02:03.871635
 revision = '00222d3a3142'
 down_revision = '9afc9a45510c'
 
-from alembic import op
+from alembic import op, context
 import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql
+from sqlalchemy.engine import reflection
+import libweasyl
 
 
 def upgrade():
     """To purge: commission, composition, contentview, ignorecontent, loginaddress, logininvite"""
     
-    # Drop 'commission'
-    op.drop_table('commission')
+    # The Weasyl latest staff DB doesn't have these tables, so alembic fails to run...
+    # Therefore: manually get/check these tables to see if they exist before operating.
+    url = context.config.get_main_option("sqlalchemy.url")
+    sqa_engine = sa.create_engine(url)
+    insp = reflection.Inspector.from_engine(sqa_engine)
+    table_names = insp.get_table_names()
+
+    # Drop 'commission' table, if it exists
+    if 'commission' in table_names:
+        op.drop_table('commission')
     
-    # Drop 'composision'
-    op.drop_table('composition')
+    # Drop 'composision' table, if it exists
+    if 'composition' in table_names:
+        op.drop_table('composition')
     
-    # Drop 'contentview'
-    op.drop_index('ind_contentview_targetid', table_name='contentview')
-    op.drop_table('contentview')
+    # Drop 'contentview' table and index, if it exists
+    if 'contentview' in table_names:
+        op.drop_index('ind_contentview_targetid', table_name='contentview')
+        op.drop_table('contentview')
     
-    # Drop 'ignorecontent'
-    op.drop_index('ind_ignorecontent_userid', table_name='ignorecontent')
-    op.drop_table('ignorecontent')
+    # Drop 'ignorecontent' table and index, if it exists
+    if 'ignorecontent' in table_names:
+        op.drop_index('ind_ignorecontent_userid', table_name='ignorecontent')
+        op.drop_table('ignorecontent')
     
-    # Drop 'loginaddress'
-    op.drop_table('loginaddress')
+    # Drop 'loginaddress' table, if it exists
+    if 'loginaddress' in table_names:
+        op.drop_table('loginaddress')
     
-    # Drop 'logininvite'
-    op.drop_table('logininvite')
+    # Drop 'logininvite' table, if it exists
+    if 'logininvite' in table_names:
+        op.drop_table('logininvite')
 
 def downgrade():
     """To create: commission, composition, contentview, ignorecontent, loginaddress, logininvite"""
